requirements.txt
flask
sqlalchemy
python-dotenv
requests
flask-cors
PyJWT
